Moved warehouse api into warehouse module
