+++ apps/web-giddh/src/app/models/api-models/Warehouse.ts
@@ -1,0 +1,12 @@
+export class WareHouseResponse {
+    countryCode: string;
+    currencyCode: string;
+    stateCode: string;
+    uniqueName: string;
+    taxType: string;
+    taxNumber: string;
+    callingCode: string;
+    mobileNumber: string;
+    address: string;
+    name: string;
+}
